# Conflicts:
#	runelite-client/src/main/java/net/runelite/client/plugins/grounditems/GroundItemsOverlay.java

+++ runelite-client/src/main/java/net/runelite/client/plugins/dpscounter/DpsMember.java
@@ -62,7 +62,7 @@ class DpsMember
 			return 0;
 		}
 
-		return (float) damage / (float) diff;
+		return damage / (float) diff;
 	}
 
 	void pause()
